requirements.txt
langchain==0.3.25
langchain-groq==0.3.2
groq==0.24.0
python-dotenv==1.1.0